requirements.txt
torch==2.1.1
torchvision==0.16.1
hydra-core
diffusers
torchaudio
tqdm
submitit

